requirements.txt
torchvision
pillow==6.2.1
pyyaml
python-lmdb
pyarrow==0.11.1
tqdm
numpy
pytorch>=1.2.0
ffmpeg==4.0
